requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn>=21.0
psycopg2-binary>=2.9
